Don't assume that id_token is present
Some providers (eg. Gitea) don't provide an id_token to cache the user info and that causes a panic.

This change checks to see if the key value is not nil before converting to a string.

With this change, Gitea can be used as a provider using either "github" or "oidc" as the configured provider.