add support for multiple login states in different tabs
the session cookie that contains the state now has its path attribute set to /auth/[state]/. /auth redirects the user to this new endpoint, which will only receive the session that was set with this specific state. /auth/[state]/ will then perform the known authentication that was formerly done by /auth. This allows for multiple login tabs to be opened at the same time with all of them still working.