embed the static files into the vouch binary
the embed package is new to Go 1.6
https://golang.org/pkg/embed/

with this change it's no longer neccesseary to distribute the static
files with vouch.

all urls under the /static/ path are served from the embeded static
filesystem. the static filesystem has all the files with the /static/
subdirectory so no stripping is needed.

also removed the copy of static files to the Docker image