introduce explicit cfg.StateDir
for now, only used by the cfg.secretFile, which is $StateDir/secret

cfg.StateDir is initialised by the environment variable STATE_DIRECTORY
compatible with what's provided by systemd (ie. the string is split on
":" and the first item is used).

when STATE_DIRECTORY is not provided, cfg.StateDir is set to $RootDir/config
to preserve the previous behavior (secretFile was
$RootDir/config/secret)